+++ pkg/frontend/util.go
@@ -632,68 +632,3 @@ func isInvalidConfigInput(config string) bool {
 	_, err := types.ParseStringToByteJson(config)
 	return err != nil
 }
-
-func removePrefixComment(sql string) string {
-	if len(sql) >= 4 {
-		p1 := strings.Index(sql, "/*")
-		if p1 != 0 {
-			// no prefix comment in this sql
-			return sql
-		}
-
-		p2 := strings.Index(sql, "*/")
-		if p2 < 2 {
-			// no valid prefix comment in this sql
-			return sql
-		}
-
-		sql = sql[p2+2:]
-	}
-	return sql
-}
-
-func splitBySemicolon(sql string) []string {
-	lastEnd := 0
-	sqlLen := len(sql)
-	ret := make([]string, 0)
-	for i := 0; i < sqlLen; i++ {
-		// skip double quotes
-		if i < sqlLen-1 && sql[i] == '"' {
-			i++
-			for i < sqlLen && sql[i] != '"' {
-				i++
-			}
-			i++
-		}
-
-		// skip single quotes
-		if i < sqlLen-1 && sql[i] == '\'' {
-			i++
-			for i < sqlLen && sql[i] != '\'' {
-				i++
-			}
-			i++
-		}
-
-		// skip comments
-		if i < sqlLen-3 && sql[i] == '/' && sql[i+1] == '*' {
-			i += 2
-			for i < sqlLen-1 && !(sql[i] == '*' && sql[i+1] == '/') {
-				i++
-			}
-			i += 2
-		}
-
-		if i < sqlLen && sql[i] == ';' {
-			ret = append(ret, sql[lastEnd:i])
-			i++
-			lastEnd = i
-		}
-	}
-
-	if lastEnd != sqlLen {
-		ret = append(ret, sql[lastEnd:])
-	}
-
-	return ret
-}
